Add files via upload
This project is a donation portal for an NGO, built using HTML, CSS, Python, and Flask. It allows users to view donation options and submit their details through a simple web interface. Flask handles the backend logic, while HTML and CSS are used to design the frontend. Submitted data is processed and can be stored or emailed for further action by the NGO.